requirements.txt
matplotlib~=3.5.2

pytesseract~=0.3.9
Pillow~=9.1.1